+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.0.2 (2015/08/13)
+
+Enchancements:
+
+* Allow to configure timeout and lock_path in config file (options outcome)
+* Allow to configure log_path and log_level as command options (options outcome)
+
 # 0.0.1 (2015/08/13)
 
 first version